requirements.txt
appdirs==1.4.3
backcall==0.1.0
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.1
dlib==19.19.0
ffmpeg-python==0.2.0
future==0.18.2
hmmlearn==0.2.3
imageio==2.6.1
imutils==0.5.3
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
joblib==0.14.1
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
matplotlib==3.1.2
networkx==2.4
numpy==1.18.1
opencv-contrib-python==4.1.2.30
pandas==1.0.1
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pyzmq==18.1.1
rply==0.7.7
scikit-image==0.16.2
scikit-learn==0.22.2
scipy==1.4.1
six==1.14.0
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.8
